FoodRecipe Code

+++ FoodRecipe/src/main/java/com/java/recipe/guid/crudoperations/repository/RecipeRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import com.java.recipe.guid.crudoperations.model.RecipeEntity;
 
 @Repository
-	public interface RecipeRepository extends JpaRepository<RecipeEntity, Long>
+	public interface RecipeRepository extends JpaRepository<RecipeEntity, Integer>
 	{
 
 	}
